+++ spec/rubocop/cops/style/module_function_spec.rb
@@ -1,0 +1,30 @@
+# encoding: utf-8
+
+require 'spec_helper'
+
+module Rubocop
+  module Cop
+    module Style
+      describe ModuleFunction do
+        let(:cop) { described_class.new }
+
+        it 'registers an offence for extend self in module' do
+          inspect_source(cop,
+                         ['module Test',
+                          '  extend self',
+                          '  def test; end',
+                          'end'])
+          expect(cop.offences.size).to eq(1)
+        end
+
+        it 'accepts extend self in class' do
+          inspect_source(cop,
+                         ['class Test',
+                          '  extend self',
+                          'end'])
+          expect(cop.offences).to be_empty
+        end
+      end
+    end
+  end
+end