+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * New cop `UselessComparison` checks for comparisons of the same arguments.
 * New cop `UselessAssignment` checks for useless assignments to local variables.
 * New cop `SignalException` checks for proper usage of `fail` and `raise`.
+* New cop `ModuleFunction` checks for usage of `extend self` in modules.
 
 ### Bugs fixed